Correct the problem of saving pictures in a trick
